Create Project 2 "Coffee"
Project Outline 

Everett Enterprises is a marketing agency for major social media companies. They currently host their business in a on-premises data center, but have decided to move to the cloud due to the financial and performance improvements. They have a website that is utilized by their clients, and it is urgent that they have a working environment for their business requirements.  

You have been hired to assist their company with the transition. You must design a fast, scalable, reliable, and cost-effective solution to replace their on-premises infrastructure.  

For this project, you will build a 3-tier web architecture in AWS. This architecture will consist of a public tier to load balance traffic to the private tiers. Private tiers will consist of the web-servers hosting the application and a backend database tier that is fully managed and highly available. 

###
~Demonstrate a working project: deploy a virtual network across multiple availability zones in a region using Terraform
~Deploy a highly available and fully managed relational database using RDS 
~Create a database caching layer using ElastiCache for frequently run queries, to improve HTTP response time performance and reduce strain on the database instantiated in the previous step 
~Create a group of web servers that will automatically scale in response to load variations to complete your application tier 
~Create a visual presentation to explain your project; include a visual aid such a Lucid Chart 
###