Project 1 Submittal
Project Outline

For this project, you are deploying a highly available company intranet web portal across multiple availability zones and EC2 instances behind a load balancer. The EC2 instances will be placed within an autoscaling group that will monitor the EC2 CPU usage as the metric to determine when to scale. They will additionally be monitored by CloudWatch and have a dashboard to display metrics and keep an eye on the deployment, as well as send SNS alerts when autoscaling will occur.

The EC2 instances will be bootstrapped and use a custom script to automatically download the websites file(s) from S3 (a role will be required to give read-only access). This script will need to automatically update, install httpd, do any required config, and display the web page while also displaying the EC2 instances instanceID and availability zone which will be retrieved from the metadata endpoint.

The goal is to gain bootstrapping experience, and deploy these resources in a secure, highly available deployment across 2 availability zones.