Project Description
Everett Enterprises is a marketing agency for major social media companies. They currently host their business in a on-premises data center, but have decided to move to the cloud due to the financial and performance improvements. They have a website that is utilized by their clients, and it is urgent that they have a working environment for their business requirements.  

You have been hired to assist their company with the transition. You must design a fast, scalable, reliable, and cost-effective solution to replace their on-premises infrastructure.  

For this project, you will build a 3-tier web architecture in AWS. This architecture will consist of a public tier to load balance traffic to the private tiers. Private tiers will consist of the web-servers hosting the application and a backend database tier that is fully managed and highly available.